Began implementing onscreen keyboard that displays colors to represent correct letters
